+++ src/games/snake.cpp
@@ -8,11 +8,11 @@ SnakeGame::SnakeGame() {
 SnakeGame::~SnakeGame() {}
 
 void SnakeGame::init() {
-    constexpr int SNAKE_START_X = 14;
-    constexpr int SNAKE_START_Y = 14;
-    constexpr int SNAKE_START_LEN = 5;
-    constexpr int SNAKE_START_COLOR = 90;
-    constexpr int START_DELAY = 300;
+    constexpr uint8_t SNAKE_START_X = 14;
+    constexpr uint8_t SNAKE_START_Y = 14;
+    constexpr uint8_t SNAKE_START_LEN = 5;
+    constexpr uint8_t SNAKE_START_COLOR = 90;
+    constexpr uint16 START_DELAY = 300;
 
     FastLED.clear(true);
     srand(millis());
@@ -47,7 +47,7 @@ void SnakeGame::update(const ControllerInput &input) {
         FastLED.clear(false);
         leds[convertXY(apple_x, apple_y)] = CHSV(255, 255, 255);
 
-        for (int i = 1; i < snake_len; ++i) {
+        for (uint16_t i = 1; i < snake_len; ++i) {
             if (snake_x[i] != -1 && snake_y[i] != -1) {
                 leds[convertXY(snake_x[i], snake_y[i])] = CHSV(snake_color, 255, 255);
             }
@@ -58,10 +58,10 @@ void SnakeGame::update(const ControllerInput &input) {
 }
 
 void SnakeGame::checkController(const ControllerInput &input) {
-    constexpr int JOY_X_LOW = 50;
-    constexpr int JOY_X_HIGH = 200;
-    constexpr int JOY_Y_LOW = 50;
-    constexpr int JOY_Y_HIGH = 200;
+    constexpr uint8_t JOY_X_LOW = 50;
+    constexpr uint8_t JOY_X_HIGH = 200;
+    constexpr uint8_t JOY_Y_LOW = 50;
+    constexpr uint8_t JOY_Y_HIGH = 200;
 
     if (input.joy_x < JOY_X_LOW && last_dir != RIGHT)
         snake_dir = LEFT;
@@ -74,7 +74,7 @@ void SnakeGame::checkController(const ControllerInput &input) {
 }
 
 void SnakeGame::moveSnake() {
-    for (int i = snake_len - 1; i > 0; i--) {
+    for (uint16_t i = snake_len - 1; i > 0; i--) {
         snake_x[i] = snake_x[i - 1];
         snake_y[i] = snake_y[i - 1];
     }
@@ -119,11 +119,11 @@ void SnakeGame::checkSnake() {
 }
 
 void SnakeGame::progressSnakeLevel() {
-    constexpr std::array<int, 5> DELAY_ARR = {200, 150, 100, 50, 35};
-    constexpr std::array<int, 5> COLOR_ARR = {120, 140, 160, 180, 200};
-    constexpr std::array<int, 5> LEN_ARR = {10, 15, 20, 25, 30};
-    constexpr int RAINBOW_SNAKE_LEN = 35;
-    constexpr int RAINBOW_DELAY = 25;
+    constexpr std::array<uint16_t, 5> DELAY_ARR = {200, 150, 100, 50, 35};
+    constexpr std::array<uint16_t, 5> COLOR_ARR = {120, 140, 160, 180, 200};
+    constexpr std::array<uint16_t, 5> LEN_ARR = {10, 15, 20, 25, 30};
+    constexpr uint8_t RAINBOW_SNAKE_LEN = 35;
+    constexpr uint8_t RAINBOW_DELAY = 25;
 
     // assert(RAINBOW_SNAKE_LEN > LEN_ARR.back());
     // assert(RAINBOW_DELAY < DELAY_ARR.back());
@@ -140,15 +140,15 @@ void SnakeGame::progressSnakeLevel() {
 
 void SnakeGame::respawnApple() {
     memset(grid, 0, sizeof(grid));
-    for (int i = 0; i < snake_len; ++i) {
+    for (uint16_t i = 0; i < snake_len; ++i) {
         if (snake_x[i] >= 0 && snake_y[i] >= 0) {
             grid[snake_x[i]][snake_y[i]] = true;
         }
     }
 
-    int free_count = 0;
-    for (int x = 0; x < WIDTH; ++x) {
-        for (int y = 0; y < HEIGHT; ++y) {
+    uint8_t free_count = 0;
+    for (uint8_t x = 0; x < WIDTH; ++x) {
+        for (uint8_t y = 0; y < HEIGHT; ++y) {
             if (!grid[x][y]) {
                 free_positions[free_count++] = {x, y};
             }
@@ -156,7 +156,7 @@ void SnakeGame::respawnApple() {
     }
 
     if (free_count > 0) {
-        int index = rand() % free_count;
+        uint16_t index = rand() % free_count;
         apple_x = free_positions[index].first;
         apple_y = free_positions[index].second;
     }
